Fix failing tests
Introduced a dedicated IntegrationTestConfig for service / data access
layer tests which excludes the JavaFX UI components.

Previously the tests failed because the SpringBoot test environment
tried to instantiate the UI application components within a test run
in a different order than when run as a standalone app.
In that case the JavaFX runtime was not bootstrapped yet when spring
tried to construct UI component instances discovered via
component-scanning.

I also introduced a JUnit 4 rule to mark test classes which require
the JavaFX environment to initialized.

This is just a workaround - a better solution would be to split
the application into two modules:
a *javafx-client*-module with the javafx frontend and a *app*-module.
The app module would contain your core business logic with services
and repositories.
The client module would then just use the service module.
Client and Business Logic could then be tested independently.
For the app-module you could have a SpringBoot bootstrap in the
test package just for that layer.